数据类型修改

+++ Backstage/citi/src/main/java/citi/vo/StrategyDAO.java
@@ -3,11 +3,11 @@ package citi.vo;
 public class StrategyDAO {
     private String strategyID;
     private String MerchantID;
-    private Integer full;
-    private Integer priceAfter;
-    private Integer points;
+    private Double full;
+    private Double priceAfter;
+    private Double points;
 
-    public StrategyDAO(String strategyID, String merchantID, Integer full, Integer priceAfter, Integer points) {
+    public StrategyDAO(String strategyID, String merchantID, Double full, Double priceAfter, Double points) {
         this.strategyID = strategyID;
         MerchantID = merchantID;
         this.full = full;
@@ -23,15 +23,15 @@ public class StrategyDAO {
         return MerchantID;
     }
 
-    public Integer getFull() {
+    public Double getFull() {
         return full;
     }
 
-    public Integer getPriceAfter() {
+    public Double getPriceAfter() {
         return priceAfter;
     }
 
-    public Integer getPoints() {
+    public Double getPoints() {
         return points;
     }
